fixing code

+++ firmware/include/ILI9341_functions.h
@@ -5,9 +5,9 @@
     void initializingDisplay();
     void eraseInformation(int row, int color=0);
     void drawHistogram(const float* data, int numBars);
-    void setChannelValue(long value1, char unit[6], int channel);
+    void setChannelValue(double value1, char unit[6], int channel);
     void setRelationValue(float value1);
     void drawInformation(const char text[100], int row, int color);
-    void setGraphicalValue(bool initialized=true, std::vector<int> channel_1={0}, std::vector<int> channel_2={0});
+    void setGraphicalValue(bool initialized=true, std::vector<double> channel_1={0}, std::vector<double> channel_2={0});
 
 #endif
